requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-genai>=1.0.0
chromadb>=0.6.0
python-multipart>=0.0.18
pytest>=8.0.0
pytest-anyio>=0.0.0
httpx>=0.28.0
